packages/agents/analyst/pyproject.toml
[project]
name = "analyst"
version = "0.1.0"
description = "天気データ分析AIエージェント - 分析エージェント（Step 2）"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
    "bedrock-agentcore>=1.4",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/analyst"]
